First real upload
This is the beta of beta, early access of early access. This is the core processing part of the project. In the next step, I will add a UI into the web app so that users can filter the data to plot, for example, assessment value, lot size, etc. At the end, I will also like to add a section for basic statistics of the data set plotted.